Fix podio fields

+++ ci4/app/Libraries/PodioLib.php
@@ -138,8 +138,8 @@ class PodioLib {
 
     public function moveStoryToTest(PodioItem $item): void {
         PodioItem::update_values($this->client, $item->item_id, [
-            getenv('PODIO_APP_PHASE_FIELD_ID') => getenv('PODIO_APP_PHASE_NEW_VALUE'),
-            getenv('PODIO_APP_PROGRESS_FIELD_ID') => getenv('PODIO_APP_PROGRESS_NEW_VALUE'),
+            getenv('PODIO_APP_PHASE_FIELD_ID') => (int)getenv('PODIO_APP_PHASE_NEW_VALUE'),
+            getenv('PODIO_APP_PROGRESS_FIELD_ID') => (int)getenv('PODIO_APP_PROGRESS_NEW_VALUE'),
         ]);
     }
